requirements.txt
Flask==3.1.0
tensorflow==2.18.0
keras==3.8.0
matplotlib==3.10.0
numpy==1.26.4
pillow==10.4.0
opencv-python==4.10.0.84